% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcensusdata.R
\name{census_data_API}
\alias{census_data_API}
\title{Census Data Api}
\usage{
census_data_API(base_url = "http://api.census.gov/data/2014/acs5?",
  state = 48, key = "6ee9b8141913fdd7763ff46af20c20d0e9a5bc68")
}
\arguments{
\item{base_url}{The base_url for the API}

\item{state}{The state the user is simulating}

\item{key}{The key needed to use the API. Can be requested}
}
\value{
census_data A dataframe of the Census data used for simulations in this package
}
\description{
This function mines the data needed for the simulations from the US Census
}
\details{
It calls other functions in the citymodels package in order to simulate characteristics for households and for each individual
}
