% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individualfunctions.R
\name{getcitizenandlang}
\alias{getcitizenandlang}
\title{Simulate Citizenship and Language at Home}
\usage{
getcitizenandlang(state, county, tract, syntheticdataset, seed, Census_data)
}
\arguments{
\item{state}{The state the user is simulating}

\item{county}{The county the user is simulating}

\item{tract}{The tract the user is simulating}

\item{syntheticdataset}{The individual simulated so far must include english speaking ability and nativity.}

\item{seed}{The seed to use for sampling.}

\item{Census_data}{Census data to use for the simulation. Can be mined from the function census_data_API}
}
\value{
syntheticdataset The simulated dataset with the added variables for language spoken at home and citizenship.
}
\description{
This function uses data from the U.S. Census to simulate the Citizenship and Language of an individual.
}
\details{
The function uses the data from the U.S. Census on the tract level to build a probability vector for citizenship and language based on presimulated nativity and english speaking ability.
English speaking ability and nativity can be simulated using the getlangandnativity function
It then samples using the user inputed seed.
}
