% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupquartersfunctions.R
\name{gethealthinsuranceforgroupquarters}
\alias{gethealthinsuranceforgroupquarters}
\title{Simulate health insurance for people in group quarters}
\usage{
gethealthinsuranceforgroupquarters(state, county, tract, syntheticdataset, seed,
  Census_data)
}
\arguments{
\item{state}{The state the user is simulating}

\item{county}{The county the user is simulating}

\item{tract}{The tract the user is simulating}

\item{syntheticdataset}{The individual simulated must include age and disability status.}

\item{seed}{The seed to use for sampling.}

\item{Census_data}{Census data to use for the simulation. Can be mined from the function census_data_API}
}
\value{
syntheticdataset The simulated dataset with the added variable for health insurance.
}
\description{
This function uses data from the U.S. Census to sample the health insurance status for people living in group quarters.
}
\details{
The function uses the data from the U.S. Census on the tract level to build a probability vector for health insurance status based off of age and disability status.
Age can be simulated with the getsexraceandage function. Disability status cen be simulated using the getdisability function.
It then samples using the user inputed seed.
}
