% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/householdfunctions.R
\name{gethouseholdincome}
\alias{gethouseholdincome}
\title{Simulate Household Income}
\usage{
gethouseholdincome(state, county, tract, syntheticdataset, seed, Census_data)
}
\arguments{
\item{state}{The state the user is simulating}

\item{county}{The county the user is simulating}

\item{tract}{The tract the user is simulating}

\item{syntheticdataset}{The household simulated.}

\item{seed}{The seed to use for sampling.}

\item{Census_data}{Census data to use for the simulation. Can be mined from the function census_data_API}
}
\value{
syntheticdataset The simulated household with the added variable of household income.
}
\description{
This function uses data from the U.S. Census on the tract level to build a probability vector for different household incomes and sample with the user inputed seed.
}
