% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individualfunctions.R
\name{getlangandnativity}
\alias{getlangandnativity}
\title{Simulate English Speaking Ability and Nativity}
\usage{
getlangandnativity(state, county, tract, syntheticdataset, seed, Census_data)
}
\arguments{
\item{state}{The state the user is simulating}

\item{county}{The county the user is simulating}

\item{tract}{The tract the user is simulating}

\item{syntheticdataset}{The individual simulated so far must include race.}

\item{seed}{The seed to use for sampling.}

\item{Census_data}{Census data to use for the simulation. Can be mined from the function census_data_API}
}
\value{
syntheticdataset The simulated dataset with the added variables for english speaking ability and nativity.
}
\description{
This function uses data from the U.S. Census to simulate the Language and Naticity of an individual.
}
\details{
The function uses the data from the U.S. Census on the tract level to build a probability vector for lenglish speaking ability and nativity based on presimulated race.
Race can be simulated with the getsexraceandage function
It then samples using the user inputed seed.
}
