% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupquartersfunctions.R
\name{getnumberofvehiclesforgroupquarters}
\alias{getnumberofvehiclesforgroupquarters}
\title{Get number of vehicles for people in group quarters}
\usage{
getnumberofvehiclesforgroupquarters(state, county, tract, syntheticdataset,
  seed, Census_data)
}
\arguments{
\item{state}{The state the user is simulating}

\item{county}{The county the user is simulating}

\item{tract}{The tract the user is simulating}

\item{syntheticdataset}{The individual simulated.}

\item{seed}{The seed to use for sampling.}

\item{Census_data}{Census data to use for the simulation. Can be mined from the function census_data_API}
}
\value{
syntheticdataset The simulated dataset with the added variable for number.of.vehicles.
}
\description{
This function uses data from the U.S. Census to sample the number of vehicles for people living in group quarters.
}
\details{
The function uses the data from the U.S. Census on the tract level to build a probability vector for the number of cars based off of sex.
Sex can be simulated using the getsexraceandage function.
It then samples using the user inputed seed.
}
