% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/householdfunctions.R
\name{getnumberofvehiclesforhouseholds}
\alias{getnumberofvehiclesforhouseholds}
\title{Get number of vehicles for households}
\usage{
getnumberofvehiclesforhouseholds(state, county, tract, syntheticdataset, seed,
  Census_data)
}
\arguments{
\item{state}{The state the user is simulating}

\item{county}{The county the user is simulating}

\item{tract}{The tract the user is simulating}

\item{syntheticdataset}{The household simulated.}

\item{seed}{The seed to use for sampling.}

\item{Census_data}{Census data to use for the simulation. Can be mined from the function census_data_API}
}
\value{
syntheticdataset The simulated household with the added variable for number of vehicles.
}
\description{
This function uses data from the U.S. Census to sample the number of vehicles a household would have based on its size.
}
\details{
The function uses the data from the U.S. Census on the tract level to build a probability vector from the number of households of the inputed size with either no vehicles, 1 vehicle, 2 vehicles, 3 vehicles or 4 or more vehicles.
It then samples using the user inputed seed.
}
