% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individualfunctions.R
\name{getsexraceandage}
\alias{getsexraceandage}
\title{Get Sex Age and Race}
\usage{
getsexraceandage(state, county, tract, syntheticdataset, seed, Census_data)
}
\arguments{
\item{state}{The state the user is simulating}

\item{county}{The county the user is simulating}

\item{tract}{The tract the user is simulating}

\item{syntheticdataset}{The individual simulated so far.}

\item{seed}{The seed to use for sampling.}

\item{Census_data}{Census data to use for the simulation. Can be mined from the function census_data_API}
}
\value{
syntheticdataset The simulated dataset with the added variables for sex, age and race.
}
\description{
This function uses data from the U.S. Census to simulate the sex, race and age of an individual.
}
\details{
The function uses the data from the U.S. Census on the tract level to build a probability vector for sex age and race which are cross tabulated together.
It then samples using the user inputed seed.
}
