% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individualfunctions.R
\name{gettraveltime}
\alias{gettraveltime}
\title{Get Transportation Time}
\usage{
gettraveltime(state, county, tract, syntheticdataset, seed, Census_data)
}
\arguments{
\item{state}{The state the user is simulating}

\item{county}{The county the user is simulating}

\item{tract}{The tract the user is simulating}

\item{syntheticdataset}{The individual simulated so far must include sex.}

\item{seed}{The seed to use for sampling.}

\item{Census_data}{Census data to use for the simulation. Can be mined from the function census_data_API}
}
\value{
syntheticdataset The simulated dataset with the added variable for transport.
}
\description{
This function uses data from the U.S. Census to simulate the time taken to get to work for an individual.
}
\details{
The function uses the data from the U.S. Census on the tract level to build a probability vector for time taken to get to work based off of mode of transport.
Sex can be simulated using the gettransport function
It then samples using the user inputed seed.
}
