% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupquarterssimulator.R
\name{group_quarters_simulater}
\alias{group_quarters_simulater}
\title{Group Quarters Simulator}
\usage{
group_quarters_simulater(state, county, tract, seed, inputdir = "../Inputs/",
  Census_data)
}
\arguments{
\item{state}{The state the user is simulating}

\item{county}{The county the user is simulating}

\item{tract}{The tract the user is simulating}

\item{seed}{The seed to use for sampling.}

\item{inputdir}{The input directory for other data}

\item{Census_data}{Census data to use for the simulation. Can be mined from the function census_data_API}
}
\value{
syntheticdataset A dataframe of simulated people living in group quarters.
}
\description{
This function simulates people living in group quarters.
}
\details{
It calls other functions in the citymodels package in order to simulate characteristics for each individual
}
