% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/householdgenerator.R
\name{household_generator}
\alias{household_generator}
\title{Household Generator}
\usage{
household_generator(state, county, tract, seed, inputdir = "../Inputs/",
  Census_data)
}
\arguments{
\item{state}{The state the user is simulating}

\item{county}{The county the user is simulating}

\item{tract}{The tract the user is simulating}

\item{seed}{The seed to use for sampling.}

\item{inputdir}{The input directory for other data}

\item{Census_data}{Census data to use for the simulation. Can be mined from the function census_data_API}

\item{number.of.people}{The number of people to simulate}
}
\value{
syntheticdataset A dataframe of simulated people living in households.
}
\description{
This function simulates people living in households.
}
\details{
It calls other functions in the citymodels package in order to simulate characteristics for households and for each individual
}
